Add Collision() to engine and create objects in engine
The obstacles now render but the collision methods are not working as exprected.

+++ Bounce/Bounce/Bounce/LineHorizontal.cs
@@ -10,7 +10,7 @@ namespace Bounce
     internal class LineHorizontal : Line
     {
         Pen Pen = new Pen(Color.Green);
-        public LineHorizontal(int x, int y) : base(x, y)
+        public LineHorizontal(float x, float y) : base(x, y)
         {
             CreateObject();
         }
